Add conformance_server.py for running server conformance tests
